:wrench: refactor(tenant): implement dynamic tenant use case configuration
- Add updateTenantUseCaseConfiguration lambda for updating tenant configs
- Replace TenantUseCaseManagement with DynamicRouter component
- Update admin operations to handle tenant-specific use case toggles
- Enhance hooks to support dynamic configuration management
- Improve admin portal routing and tenant registration flow
