+++ packages/web/src/hooks/useTenantUseCaseConfig.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { TenantUseCaseConfigResponse } from 'generative-ai-use-cases-types';
+import { TenantUseCaseConfigResponse } from 'generative-ai-use-cases';
 import useHttp from './useHttp';
 
 const useTenantUseCaseConfig = () => {
@@ -21,17 +21,7 @@ const useTenantUseCaseConfig = () => {
     } catch (err) {
       console.error('Failed to fetch tenant use case configuration:', err);
       setError(err instanceof Error ? err.message : 'Failed to fetch configuration');
-      
-      // Fallback to global configuration from environment variables
-      const globalConfig: HiddenUseCases = JSON.parse(
-        import.meta.env.VITE_APP_HIDDEN_USE_CASES || '{}'
-      );
-      
-      setTenantConfig({
-        tenantId: null,
-        hiddenUseCases: globalConfig,
-        source: 'global_fallback',
-      });
+      setTenantConfig(null);
     } finally {
       setLoading(false);
     }
