+++ packages/web/src/hooks/useUseCases.ts
@@ -1,19 +1,14 @@
-import { HiddenUseCases, HiddenUseCasesKeys } from 'generative-ai-use-cases';
+import { HiddenUseCasesKeys } from 'generative-ai-use-cases';
 import useTenantUseCaseConfig from './useTenantUseCaseConfig';
 
-// Fallback to global configuration from environment variables
-const globalHiddenUseCases: HiddenUseCases = JSON.parse(
-  import.meta.env.VITE_APP_HIDDEN_USE_CASES || '{}'
-);
-
 const useUseCases = () => {
-  const { tenantConfig } = useTenantUseCaseConfig();
-
-  // Use tenant-specific configuration if available, otherwise use global configuration
-  const hiddenUseCases = tenantConfig?.hiddenUseCases || globalHiddenUseCases;
+  const { tenantConfig, loading } = useTenantUseCaseConfig();
 
   const enabledSingle = (useCase: HiddenUseCasesKeys): boolean => {
-    return !hiddenUseCases[useCase];
+    if (!tenantConfig?.hiddenUseCases) {
+      return true; // Enable all use cases if no tenant configuration
+    }
+    return !tenantConfig.hiddenUseCases[useCase];
   };
 
   const enabled = (...useCases: HiddenUseCasesKeys[]): boolean => {
@@ -22,8 +17,8 @@ const useUseCases = () => {
 
   return {
     enabled,
-    tenantConfig, // Expose tenant config for debugging/information
-    loading: tenantConfig === null, // Still loading if no config yet
+    tenantConfig,
+    loading,
   };
 };
