+++ packages/web/src/main.tsx
@@ -1,56 +1,11 @@
 import './i18n/config';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import AuthWithUserpool from './components/AuthWithUserpool';
-import AuthWithSAML from './components/AuthWithSAML';
-import AuthWithSamlOrUserpool from './components/AuthWithSamlOrUserpool';
 import './index.css';
-import {
-  RouterProvider,
-  createBrowserRouter,
-  RouteObject,
-} from 'react-router-dom';
 import { SWRConfig } from 'swr';
-import LandingPage from './pages/LandingPage';
-import Setting from './pages/Setting';
-import StatPage from './pages/StatPage.tsx';
-import ChatPage from './pages/ChatPage';
-import SharedChatPage from './pages/SharedChatPage';
-import SummarizePage from './pages/SummarizePage';
-import GenerateTextPage from './pages/GenerateTextPage';
-import TranslatePage from './pages/TranslatePage';
-import VideoAnalyzerPage from './pages/VideoAnalyzerPage';
-import NotFound from './pages/NotFound';
-import RagPage from './pages/RagPage';
-import RagKnowledgeBasePage from './pages/RagKnowledgeBasePage';
-import WebContent from './pages/WebContent';
-import GenerateImagePage from './pages/GenerateImagePage';
-import GenerateVideoPage from './pages/GenerateVideoPage';
-import OptimizePromptPage from './pages/OptimizePromptPage';
-import TranscribePage from './pages/TranscribePage';
-import MeetingMinutesPage from './pages/MeetingMinutesPage';
-import AgentChatPage from './pages/AgentChatPage.tsx';
-import FlowChatPage from './pages/FlowChatPage';
-import VoiceChatPage from './pages/VoiceChatPage';
-import McpChatPage from './pages/McpChatPage';
-import { MODELS } from './hooks/useModel';
 import { Authenticator } from '@aws-amplify/ui-react';
-import UseCaseBuilderEditPage from './pages/useCaseBuilder/UseCaseBuilderEditPage.tsx';
-import App from './App.tsx';
-import UseCaseBuilderRoot from './UseCaseBuilderRoot.tsx';
-import UseCaseBuilderExecutePage from './pages/useCaseBuilder/UseCaseBuilderExecutePage.tsx';
-import UseCaseBuilderSamplesPage from './pages/useCaseBuilder/UseCaseBuilderSamplesPage.tsx';
-import UseCaseBuilderMyUseCasePage from './pages/useCaseBuilder/UseCaseBuilderMyUseCasePage.tsx';
-import { optimizePromptEnabled } from './hooks/useOptimizePrompt';
-import GenerateDiagramPage from './pages/GenerateDiagramPage.tsx';
-import WriterPage from './pages/WriterPage.tsx';
-import RagChatBotPage from './pages/RagChatBotPage.tsx';
-import RagChatBotEditPage from './pages/RagChatBotEditPage.tsx';
-import RagChatBotChatPage from './pages/RagChatBotChatPage.tsx';
-import RagChatBotHistoryPage from './pages/RagChatBotHistoryPage.tsx';
-import AdminPortal from './pages/AdminPortal.tsx';
-import useUseCases from './hooks/useUseCases';
 import { Toaster } from 'sonner';
+import DynamicRouter from './components/DynamicRouter';
 
 const ragEnabled: boolean = import.meta.env.VITE_APP_RAG_ENABLED === 'true';
 const ragKnowledgeBaseEnabled: boolean =
@@ -62,261 +17,8 @@ const samlDefaultAuthEnabled: boolean =
 const agentEnabled: boolean = import.meta.env.VITE_APP_AGENT_ENABLED === 'true';
 const inlineAgents: boolean = import.meta.env.VITE_APP_INLINE_AGENTS === 'true';
 const mcpEnabled: boolean = import.meta.env.VITE_APP_MCP_ENABLED === 'true';
-const {
-  visionEnabled,
-  imageGenModelIds,
-  videoGenModelIds,
-  speechToSpeechModelIds,
-} = MODELS;
 const useCaseBuilderEnabled: boolean =
   import.meta.env.VITE_APP_USE_CASE_BUILDER_ENABLED === 'true';
-// eslint-disable-next-line  react-hooks/rules-of-hooks
-const { enabled } = useUseCases();
-
-const routes: RouteObject[] = [
-  {
-    path: '/',
-    element: <LandingPage />,
-  },
-  {
-    path: '/setting',
-    element: <Setting />,
-  },
-  {
-    path: '/stats',
-    element: <StatPage />,
-  },
-  {
-    path: '/chat',
-    element: <ChatPage />,
-  },
-  {
-    path: '/chat/:chatId',
-    element: <ChatPage />,
-  },
-  {
-    path: '/share/:shareId',
-    element: <SharedChatPage />,
-  },
-  enabled('generate')
-    ? {
-        path: '/generate',
-        element: <GenerateTextPage />,
-      }
-    : null,
-  enabled('summarize')
-    ? {
-        path: '/summarize',
-        element: <SummarizePage />,
-      }
-    : null,
-  enabled('meetingMinutes')
-    ? {
-        path: '/meeting-minutes',
-        element: <MeetingMinutesPage />,
-      }
-    : null,
-  enabled('writer')
-    ? {
-        path: '/writer',
-        element: <WriterPage />,
-      }
-    : null,
-  enabled('translate')
-    ? {
-        path: '/translate',
-        element: <TranslatePage />,
-      }
-    : null,
-  enabled('webContent')
-    ? {
-        path: '/web-content',
-        element: <WebContent />,
-      }
-    : null,
-  imageGenModelIds.length > 0 && enabled('image')
-    ? {
-        path: '/image',
-        element: <GenerateImagePage />,
-      }
-    : null,
-  videoGenModelIds.length > 0 && enabled('video')
-    ? {
-        path: '/video',
-        element: <GenerateVideoPage />,
-      }
-    : null,
-  enabled('diagram')
-    ? {
-        path: '/diagram',
-        element: <GenerateDiagramPage />,
-      }
-    : null,
-  optimizePromptEnabled
-    ? {
-        path: '/optimize',
-        element: <OptimizePromptPage />,
-      }
-    : null,
-  {
-    path: '/transcribe',
-    element: <TranscribePage />,
-  },
-  {
-    path: '/flow-chat',
-    element: <FlowChatPage />,
-  },
-  visionEnabled && enabled('videoAnalyzer')
-    ? {
-        path: '/video-analyzer',
-        element: <VideoAnalyzerPage />,
-      }
-    : null,
-  ragEnabled
-    ? {
-        path: '/rag',
-        element: <RagPage />,
-      }
-    : null,
-  ragKnowledgeBaseEnabled
-    ? {
-        path: '/rag-knowledge-base',
-        element: <RagKnowledgeBasePage />,
-      }
-    : null,
-  agentEnabled && !inlineAgents
-    ? {
-        path: '/agent',
-        element: <AgentChatPage />,
-      }
-    : null,
-  agentEnabled && inlineAgents
-    ? {
-        path: '/agent/:agentName',
-        element: <AgentChatPage />,
-      }
-    : null,
-  speechToSpeechModelIds.length > 0 && enabled('voiceChat')
-    ? {
-        path: '/voice-chat',
-        element: <VoiceChatPage />,
-      }
-    : null,
-  mcpEnabled
-    ? {
-        path: '/mcp',
-        element: <McpChatPage />,
-      }
-    : null,
-  {
-    path: '/rag-chat-bot',
-    element: <RagChatBotPage />,
-  },
-  {
-    path: '/rag-chat-bot/create',
-    element: <RagChatBotEditPage />,
-  },
-  {
-    path: '/rag-chat-bot/edit/:botId',
-    element: <RagChatBotEditPage />,
-  },
-  {
-    path: '/rag-chat-bot/chat/:botId',
-    element: <RagChatBotChatPage />,
-  },
-  {
-    path: '/rag-chat-bot/chat/:botId/:conversationId',
-    element: <RagChatBotChatPage />,
-  },
-  {
-    path: '/rag-chat-bot/history',
-    element: <RagChatBotHistoryPage />,
-  },
-  {
-    path: '/admin',
-    element: <AdminPortal />,
-  },
-  {
-    path: '*',
-    element: <NotFound />,
-  },
-].flatMap((r) => (r !== null ? [r] : []));
-
-const useCaseBuilderRoutes: RouteObject[] = [
-  {
-    path: '/use-case-builder',
-    element: <UseCaseBuilderSamplesPage />,
-  },
-  {
-    path: `/use-case-builder/my-use-case`,
-    element: <UseCaseBuilderMyUseCasePage />,
-  },
-  {
-    path: `/use-case-builder/new`,
-    element: <UseCaseBuilderEditPage />,
-  },
-  {
-    path: `/use-case-builder/edit/:useCaseId`,
-    element: <UseCaseBuilderEditPage />,
-  },
-  {
-    path: `/use-case-builder/execute/:useCaseId`,
-    element: <UseCaseBuilderExecutePage />,
-  },
-  {
-    path: `/use-case-builder/setting`,
-    element: <Setting />,
-  },
-  {
-    path: '*',
-    element: <NotFound />,
-  },
-].flatMap((r) => (r !== null ? [r] : []));
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: samlAuthEnabled ? (
-      samlDefaultAuthEnabled ? (
-        <AuthWithSamlOrUserpool>
-          <App />
-        </AuthWithSamlOrUserpool>
-      ) : (
-        <AuthWithSAML>
-          <App />
-        </AuthWithSAML>
-      )
-    ) : (
-      <AuthWithUserpool>
-        <App />
-      </AuthWithUserpool>
-    ),
-    children: routes,
-  },
-  ...(useCaseBuilderEnabled
-    ? [
-        {
-          path: '/use-case-builder',
-          element: samlAuthEnabled ? (
-            samlDefaultAuthEnabled ? (
-              <AuthWithSamlOrUserpool>
-                <UseCaseBuilderRoot />
-              </AuthWithSamlOrUserpool>
-            ) : (
-              <AuthWithSAML>
-                <UseCaseBuilderRoot />
-              </AuthWithSAML>
-            )
-          ) : (
-            <AuthWithUserpool>
-              <UseCaseBuilderRoot />
-            </AuthWithUserpool>
-          ),
-          children: useCaseBuilderRoutes,
-        },
-      ]
-    : []),
-]);
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
@@ -329,7 +31,16 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
             revalidateOnReconnect: true,
             revalidateOnMount: true,
           }}>
-          <RouterProvider router={router} />
+          <DynamicRouter
+            ragEnabled={ragEnabled}
+            ragKnowledgeBaseEnabled={ragKnowledgeBaseEnabled}
+            samlAuthEnabled={samlAuthEnabled}
+            samlDefaultAuthEnabled={samlDefaultAuthEnabled}
+            agentEnabled={agentEnabled}
+            inlineAgents={inlineAgents}
+            mcpEnabled={mcpEnabled}
+            useCaseBuilderEnabled={useCaseBuilderEnabled}
+          />
         </SWRConfig>
         <Toaster />
       </Authenticator.Provider>
